projects/mathlock-play/backend/requirements.txt
# MathLock Play — Python Dependencies
# Last audited: 2026-05-02 (pip-audit: clean)
# Install: pip install -r requirements.txt
# Audit:   pip-audit -r requirements.txt

Django==4.2.30
djangorestframework==3.17.1
django-cors-headers==4.7.0
gunicorn==25.3.0
google-api-python-client==2.194.0
google-auth==2.49.2
psycopg2-binary==2.9.11
celery[redis]==5.6.3
redis==6.4.0
